# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Lion Roaring'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
  /api/v1/login:
    post:
      summary: Login
      operationId: login
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n\"token\": \"dsdsdsd\"\n\"status\": true\n\"message\": \"Login successful\"\n}"
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_name:
                  type: string
                  description: 'The username or email of the user.'
                  example: john_doe
                password:
                  type: string
                  description: 'The password of the user.'
                  example: password
              required:
                - user_name
                - password
      security: []
...
  /api/v1/user/update-profile:
    post:
      summary: 'Update Profile'
      operationId: updateProfile
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: 'Profile updated successfully'
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Profile updated successfully'
      tags:
        - Profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  description: 'The first name of the user.'
                  example: John
                middle_name:
                  type: string
                  description: 'optional The middle name of the user.'
                  example: Doe
                last_name:
                  type: string
                  description: 'The last name of the user.'
                  example: Doe
                phone:
                  type: string
                  description: 'The phone number of the user.'
                  example: '7415236986'
                address:
                  type: string
                  description: 'The address of the user.'
                  example: '51 DN Block Merlin Infinite Building, 9th Floor, Unit, 907, Sector V, Bidhannagar, Kolkata, West Bengal 700091'
              required:
                - first_name
                - last_name
                - phone
                - address
  /api/v1/user/change-password:
    post:
      summary: 'Change Password'
      operationId: changePassword
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: 'Password changed successfully'
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Password changed successfully'
      tags:
        - Profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  description: 'The old password of the user.'
                  example: password
                new_password:
                  type: string
                  description: 'The new password of the user.'
                  example: password123
                confirm_password:
                  type: string
                  description: 'The confirm password of the user.'
                  example: password123
              required:
                - old_password
                - new_password
                - confirm_password
...